[AFV] Change output format and processing to 32 bit float
32 bit float is what is used in C# reference and has a much higher
dynamic range. 16 bit integer was clipping very often with all the VHF
simulation applied.

+++ src/blackcore/afv/audio/callsignsampleprovider.h
@@ -44,7 +44,7 @@ namespace BlackCore
                 //! Ctor
                 CallsignSampleProvider(const QAudioFormat &audioFormat, const BlackCore::Afv::Audio::CReceiverSampleProvider *receiver, QObject *parent = nullptr);
 
-                int readSamples(QVector<qint16> &samples, qint64 count) override;
+                int readSamples(QVector<float> &samples, qint64 count) override;
 
                 //! The callsign
                 const QString &callsign() const { return m_callsign; }
