+++ src/blackcore/afv/audio/output.h
@@ -60,7 +60,7 @@ namespace BlackCore
 
                 static constexpr int SampleCountPerEvent = 4800;
                 QAudioFormat m_outputFormat;
-                double m_maxSampleOutput = 0;
+                float m_maxSampleOutput = 0.0;
                 int m_sampleCount        = 0;
                 const double m_maxDb     = 0;
                 const double m_minDb     = -40;
