+++ src/blacksound/sampleprovider/bufferedwaveprovider.cpp
@@ -13,7 +13,7 @@ namespace BlackSound
             m_maxBufferSize = format.bytesForDuration(10 * 1000 * 1000);
         }
 
-        void CBufferedWaveProvider::addSamples(const QVector<qint16> &samples)
+        void CBufferedWaveProvider::addSamples(const QVector<float> &samples)
         {
             int delta = m_audioBuffer.size() + samples.size() - m_maxBufferSize;
             if (delta > 0)
@@ -23,7 +23,7 @@ namespace BlackSound
             m_audioBuffer.append(samples);
         }
 
-        int CBufferedWaveProvider::readSamples(QVector<qint16> &samples, qint64 count)
+        int CBufferedWaveProvider::readSamples(QVector<float> &samples, qint64 count)
         {
             qint64 len = qMin(count, static_cast<qint64>(m_audioBuffer.size()));
             samples = m_audioBuffer.mid(0, len);
